Fix app.rb for Ruby 3.3.
Since Ruby 3.3 [0], bundled_gems.rb intercepts `require` calls with bundler.
Under sinatra, this causes issues. Sinatra searches for the `app.rb`
style files in the call stack. Once found, sinatra uses that file as a
project's root. From there it searches for other default directories such
`views` and other.

Example of such failure from logs:
~~~
2024-05-14 12:22:14 - Errno::ENOENT - No such file or directory @ rb_sysopen - /usr/share/ruby/views/main.erb
~~~
This error appears only once a request was made to the part of code that
relies on the relative default paths, such as the './views' for the erb templates.

This was fixed with Sinatra 3.2.0[1], but to keep compatibility with Ruby
2.5, Sinatra version 2.* should remain for a while as Sinatra 3.*
restricts minimum Ruby version to Ruby 2.6.

We work around it ourselves by overriding the :app_file setting in
`app.rb` file to `__FILE__` which will fix the rest of paths.

[0] https://github.com/ruby/ruby/commit/214f6d6598be8437284dede9eb9a041a3ebfb24e
[1] https://github.com/sinatra/sinatra/blob/main/CHANGELOG.md#320--2023-12-29